requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
shap
tensorflow
joblib
streamlit
imbalanced-learn
plotly
pillow
